Adds damage to actor collisions

+++ Assets/Scripts/Actors/IDamagable.cs
@@ -2,6 +2,6 @@
 {
     public interface IDamagable
     {
-        void ApplyDamage(int damage);
+        void ApplyDamage(float damage);
     }
 }
